refactor(billing): call terminology search from the client, drop the zambda
